Added cache generation, moved Users class to models.py
