+++ app.py
@@ -1,8 +1,9 @@
 # coding=utf-8
-
+import logging
 from flask import Flask, render_template, request
 
 app = Flask(__name__)
+logging.basicConfig(level=logging.INFO)
 
 
 # index.html
